Moving view errors to a new module
These errors were used by views outside of rest dispatching, and I think it makes sense to have a module that covers the error cases.  Also added a few http and myuw http status code constants.
