+++ myuw/views/api/hfs.py
@@ -6,7 +6,8 @@ from django.http import HttpResponse
 from myuw.dao.hfs import get_account_balances_for_current_user
 from myuw.logger.timer import Timer
 from myuw.logger.logresp import log_success_response
-from myuw.views.rest_dispatch import RESTDispatch, handle_exception
+from myuw.views.rest_dispatch import RESTDispatch
+from myuw.views.error import handle_exception
 
 
 logger = logging.getLogger(__name__)
