+++ myuw/views/api/library.py
@@ -5,7 +5,8 @@ from django.http import HttpResponse
 from myuw.dao.library import get_account_info_for_current_user
 from myuw.logger.timer import Timer
 from myuw.logger.logresp import log_success_response
-from myuw.views.rest_dispatch import RESTDispatch, handle_exception
+from myuw.views.rest_dispatch import RESTDispatch
+from myuw.views.error import handle_exception
 
 
 logger = logging.getLogger(__name__)
